[PATCH] ia32 NUMA: physnode_map entries can be negative
Based on work from Bill Irwin <wli@holomorphy.com>

physnode_map[] needs to be signed so that pfn_to_nid() can return negative
values used to detect invalid pfn's.

+++ arch/i386/kernel/numaq.c
@@ -92,7 +92,7 @@ static void __init initialize_physnode_map(void)
 			cur = eq->hi_shrd_mem_start;
 			topofmem = eq->hi_shrd_mem_start + eq->hi_shrd_mem_size;
 			while (cur < topofmem) {
-				physnode_map[cur >> 8] = nid;
+				physnode_map[cur >> 8] = (s8) nid;
 				cur ++;
 			}
 		}
